tests/datawizard/in_place_partition.c: fix prototype for opencl kernel, and update error checking for main application

+++ tests/datawizard/scal_opencl.cl
@@ -1,6 +1,6 @@
 /* StarPU --- Runtime system for heterogeneous multicore architectures.
  *
- * Copyright (C) 2010, 2011  Centre National de la Recherche Scientifique
+ * Copyright (C) 2010, 2011, 2012  Centre National de la Recherche Scientifique
  * Copyright (C) 2011  Université Bordeaux 1
  *
  * StarPU is free software; you can redistribute it and/or modify
@@ -15,7 +15,7 @@
  * See the GNU Lesser General Public License in COPYING.LGPL for more details.
  */
 
-__kernel void vector_mult_opencl(__global float* val, unsigned offset, unsigned nx)
+__kernel void vector_mult_opencl(__global unsigned* val, unsigned offset, unsigned nx)
 {
         const int i = get_global_id(0);
 	val = (__global void*) val + offset;
